requirements.txt
# Core
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
websockets

# Retell AI
retell-sdk

# OpenAI
openai

# Google Calendar
google-api-python-client
google-auth
google-auth-oauthlib

# Utilities
httpx
